requirements.txt
flask>=2.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
psycopg2-binary>=2.9.0